requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sqlalchemy>=2.0.30
aiomysql>=0.2.0
pydantic-settings>=2.3.0
groq>=0.9.0
greenlet>=3.0.3
openai-whisper>=20231117
transformers>=4.41.0
torch>=2.3.0
mcp>=1.0.0
aiosqlite

# Coqui TTS does not currently publish compatible builds for Python 3.14.
# Install it separately on Python 3.11/3.12 for full voice synthesis support.
TTS>=0.22.0; python_version < "3.12"
